pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "denku"
version = "0.1.4"
description = "Custom computer vision utilities for image and video processing, visualization, and memory management."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache"}
authors = [
    {name = "Karachev Denis", email = "welcomedenk@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Image Processing"
]
keywords = ["computer-vision", "image-processing", "video-processing", "visualization", "opencv", "numpy", "pytorch"]
dependencies = [
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "opencv-python>=4.5.0",
    "opencv-python-headless>=4.5.0",
    "opencv-contrib-python>=4.5.0",
    "tqdm>=4.65.0",
    "requests>=2.28.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pre-commit>=3.0.0,<4.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<7.0.0",
    "mypy>=1.0.0,<2.0.0"
]

[project.urls]
homepage = "https://github.com/TheDenk/denku"
bug-tracker = "https://github.com/TheDenk/denku/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=denku --cov-report=term-missing"

[tool.coverage.run]
source = ["denku"]
omit = ["tests/*", "setup.py"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
